refactor(ios): extract HeaderRepository domain repository for SPV header storage
