+++ ios/StableChannels/StableChannels/Services/DatabaseService.swift
@@ -16,6 +16,7 @@ class DatabaseService {
     let pendingOpRepo: PendingOperationRepository
     let onchainRepo: OnchainReceiveRepository
     let priceRepo: PriceRepository
+    let headerRepo: HeaderRepository
 
     init(dataDir: URL) throws {
         try? FileManager.default.createDirectory(at: dataDir, withIntermediateDirectories: true)
@@ -41,6 +42,7 @@ class DatabaseService {
         self.pendingOpRepo = PendingOperationRepository(rawSQL: sqlHelper)
         self.onchainRepo = OnchainReceiveRepository(rawSQL: sqlHelper)
         self.priceRepo = PriceRepository(rawSQL: sqlHelper)
+        self.headerRepo = HeaderRepository(rawSQL: sqlHelper)
 
         sqlHelper.getDB = { [weak self] in self?.db }
 
@@ -1468,82 +1470,33 @@ class DatabaseService {
 // MARK: - Block Headers (SPV)
 
 extension DatabaseService {
-    /// Maximum number of block headers to retain in the local chain.
-    /// 2016 = one Bitcoin difficulty epoch (~2 weeks): wide enough to cover any realistic
-    /// reorg depth while keeping the table under ~160 KB on disk.
-    static let headerRetentionDepth: UInt32 = 2016
+    static var headerRetentionDepth: UInt32 { HeaderRepository.headerRetentionDepth }
 
     func insertHeader(height: UInt32, hash: String, prevHash: String, timestamp: UInt32) throws {
-        let sql = """
-        INSERT OR REPLACE INTO block_headers (height, hash, prev_hash, timestamp)
-        VALUES (?, ?, ?, ?);
-        """
-        try rawSQL.execute(sql, params: [
-            .integer(Int64(height)),
-            .text(hash),
-            .text(prevHash),
-            .integer(Int64(timestamp))
-        ])
+        try headerRepo.insertHeader(height: height, hash: hash, prevHash: prevHash, timestamp: timestamp)
     }
 
-    /// Returns true if we already have a header for this height+hash (idempotent guard).
     func headerExists(height: UInt32, hash: String) throws -> Bool {
-        let sql = "SELECT 1 FROM block_headers WHERE height = ? AND hash = ? LIMIT 1;"
-        let rows = try rawSQL.query(sql, params: [.integer(Int64(height)), .text(hash)])
-        return !rows.isEmpty
+        try headerRepo.headerExists(height: height, hash: hash)
     }
 
     func fetchLatestHeader() throws -> BlockHeaderRecord? {
-        let sql = "SELECT height, hash, prev_hash, timestamp FROM block_headers ORDER BY height DESC LIMIT 1;"
-        let rows = try rawSQL.query(sql)
-        guard let row = rows.first,
-              let rawHeight = row[0] as? Int64,
-              let hash = row[1] as? String,
-              let prevHash = row[2] as? String,
-              let rawTimestamp = row[3] as? Int64,
-              let height = UInt32(exactly: rawHeight),
-              let timestamp = UInt32(exactly: rawTimestamp) else {
-            return nil
-        }
-        return BlockHeaderRecord(
-            height: height,
-            hash: hash,
-            prevHash: prevHash,
-            timestamp: timestamp
-        )
+        try headerRepo.fetchLatestHeader()
     }
 
     func findCommonAncestorHeight(prevHash: String) throws -> UInt32? {
-        let sql = "SELECT height FROM block_headers WHERE hash = ? LIMIT 1;"
-        let rows = try rawSQL.query(sql, params: [.text(prevHash)])
-        guard let row = rows.first,
-              let rawHeight = row[0] as? Int64,
-              let height = UInt32(exactly: rawHeight) else {
-            return nil
-        }
-        return height
+        try headerRepo.findCommonAncestorHeight(prevHash: prevHash)
     }
 
     func rollbackHeadersAbove(height: UInt32) throws {
-        let sql = "DELETE FROM block_headers WHERE height > ?;"
-        try rawSQL.execute(sql, params: [.integer(Int64(height))])
+        try headerRepo.rollbackHeadersAbove(height: height)
     }
 
     func rollbackPaymentsConfirmedAfter(height: UInt32) throws {
-        let sql = """
-        UPDATE payments
-        SET status = 'pending', confirmations = 0, tx_block_height = NULL
-        WHERE tx_block_height IS NOT NULL AND tx_block_height > ?;
-        """
-        try rawSQL.execute(sql, params: [.integer(Int64(height))])
+        try headerRepo.rollbackPaymentsConfirmedAfter(height: height)
     }
 
-    /// Rolling-window prune that keeps only the most recent `headerRetentionDepth` headers.
-    /// Older entries are silently discarded — they predate any realistic reorg depth.
     func pruneOldHeaders(currentHeight: UInt32) throws {
-        guard currentHeight >= Self.headerRetentionDepth else { return }
-        let cutoff = currentHeight - Self.headerRetentionDepth
-        let sql = "DELETE FROM block_headers WHERE height < ?;"
-        try rawSQL.execute(sql, params: [.integer(Int64(cutoff))])
+        try headerRepo.pruneOldHeaders(currentHeight: currentHeight)
     }
 }
